==> 0_spotify_charts.ipynb <==
# Import dependencies
import pandas as pd
import requests
import io

# Read country code data and store in dataframe
countries = pd.read_csv('Resources/country_code.csv')
countries.head()

# Initialize dictionary to store dataframes with information on top 200 songs for various countries
dataframe_dict = {}

# Assign target date/period to a variable
period = '2016-12-23--2016-12-30'

# Retrieve top 200 charts for various countries from Spotify Charts
# for i in range(0,1):  #use for testing
for i in range(len(countries)):
    
    # Assign country code for taget country (chart) to a variable and print message to indicate that request is underway
    country = countries.loc[i,'code']
    print(f"Retrieving Top 200 Chart for {countries.loc[i,'country']}...")
    
    # Assign target url for csv file to a variable.  Print url for information.
    # url = f'https://spotifycharts.com/regional/{country}/weekly/{period}/download'  #use for weekly charts
    url = f'https://spotifycharts.com/regional/{country}/daily/latest/download'  #use for daily charts
    print(f'    {url}')
    
    # Use requests.get to retrieve information from target csv
    r = requests.get(url)
    
    try:
          # Use 'utf-8' to allow for proper interpretation of retrieved data
          data = r.content.decode('utf-8')
          
          # Read and assign retrieved data to a dataframe
          chart = pd.read_csv(io.StringIO(data))
          
          # Assign resulting dataframe as an value/entry in dataframe_dict dictionary with the country name as the key
          dataframe_dict[countries.loc[i,'country']] = chart
          
          # Use first row of dataframe as column headers and delete row
          dataframe_dict[countries.loc[i,'country']].columns = dataframe_dict[countries.loc[i,'country']].iloc[0] 
          dataframe_dict[countries.loc[i,'country']] = dataframe_dict[countries.loc[i,'country']].drop(dataframe_dict[countries.loc[i,'country']].index[0])
          
          # Add a column to the dataframe to identify the country associated with the retrieved chart data
          dataframe_dict[countries.loc[i,'country']]['Chart'] = countries.loc[i,'country']
    
    except:
          print('    Error encountered...')

# Generate list of keys (i.e., dataframes of top 200 chart information for various countries)
dataframe_dict.keys()

# Consolidate top 200 chart information from various countries into one dataframe
for key in dataframe_dict:
    
    if key == 'United States':
        spotify_charts = dataframe_dict[key]
    else:
        spotify_charts = spotify_charts.append(dataframe_dict[key])

spotify_charts

# Save data as csv file 
spotify_charts.to_csv('spotify_charts.csv',index=False)
