==> 3_hooktheory_5chord.ipynb <==
# import dependencies
import json
import requests
import time
import pandas as pd
from config import username, pword

# request credentials using username and password
url = "https://api.hooktheory.com/v1/users/auth"
header = {"accept": "application/json", 
          "Content-Type": "application/json",
          "username": username,
          "password": pword}

response = requests.post(url, data=header).json()

# save the key and id from response
key = response['activkey']
user_id = response['id']

# create header used for queries
bearer = f'Bearer {key}'
a_header = {'Authorization':bearer}

c = requests.session()
c.headers = a_header

# set query url and parameters
url = "https://api.hooktheory.com/v1/trends/songs"
params = {'cp':'', 'page':1}

# list progressions we want to check
chords_df = pd.read_csv('Resources/5_chords.csv', header=None)
progressions = chords_df[0].values.tolist()
print(len(progressions))

# initialize a list to store data
songs = []

comm = '''params['cp'] = progressions[0]
r = c.get(url, params = params)
response = r.json()
r.headers['X-Rate-Limit-Reset']'''

# query HookTheory api for information on songs utilizing target chord progressions
# use sleep method to pause script to avoid exceeding rate limits
for cp in progressions:
    params['cp'] = cp
    page = 1
    while page > 0:
        params['page'] = page
        r = c.get(url, params = params)
        count = int(r.headers['X-Rate-Limit-Remaining'])
        if count == 0:
            wait = int(r.headers['X-Rate-Limit-Reset'])
            print(len(songs))
            time.sleep(wait)
        response = r.json()
        if len(response) == 0:
            page = 0
        else:
            for song in response:
                try:
                    song['cp'] = cp
                    songs.append(song)
                except:
                    pass
            page += 1

# assign retrieved data to a dataframe
songs_df = pd.DataFrame(songs)

# drop column containing url information
songs_df = songs_df.drop(columns='url')

# display first 30 entries of dataframe
songs_df.head(30)

# save dataframe as a csv file
songs_df.to_csv('5chord_songs.csv', index=False, header=True)

# display last 30 entries of dataframe
songs_df.tail(30)
