==> 4_merge_hooktheory_data.ipynb <==
# Import dependencies
import pandas as pd
import matplotlib.pyplot as plt
import time

# Import HookTheory data
chords_3 = pd.read_csv('3chord_songs.csv')
chords_4 = pd.read_csv('4chord_songs.csv')
chords_5 = pd.read_csv('5chord_songs.csv')

# Add column to dataframe to indicate data subset
chords_3['type'] = '3 chord'

# Display portion of dataframe to verify information and format
chords_3.head(1)

# Add column to dataframe to indicate data subset
chords_4['type'] = '4 chord'

# Display portion of dataframe to verify information and format
chords_4.head(1)

# Add column to dataframe to indicate data subset
chords_5['type'] = '5 chord'

# Display portion of dataframe to verify information and format
chords_5.head(1)

# Create new dataframe to store information from all imported dataframes
chords_all = chords_3.copy()
chords_all.tail(1)

# Append information from second dataframe to new dataframe
chords_all = chords_all.append(chords_4)
chords_all.tail(1)

# Append information from third dataframe to new dataframe
chords_all = chords_all.append(chords_5)
chords_all.tail(1)

# save dataframe with consolidated information as a csv file
chords_all.to_csv('chords_all_songs.csv')
