==> 5_unique_songs.ipynb <==
# Import dependencies
import pandas as pd

# Import data retrieved from HookTheory API
hook_data = pd.read_csv('chords_all_songs.csv')

# Drop the first column which duplicates the index information
hook_data = hook_data.drop(columns='Unnamed: 0')
hook_data.head()

# Sort data based on artist and song information in ascending order
hook_data = hook_data.sort_values(by=['artist','song'])
hook_data.head(50)

# Drop duplicate entries based on artist and song title to create a unique list of songs
hook_data = hook_data.drop_duplicates(subset=['artist','song'])
hook_data.head(50)

# Display size of resulting dataframe
len(hook_data)

# Save dataframe of unique songs as a csv file
hook_data.to_csv('unique_songs.csv',index=False)
