==> 8_clean_api_data.ipynb <==
# Import dependencies
import pandas as pd
import matplotlib.pyplot as plt
import time

# Import data retrieved from HookTheory API
hook_data = pd.read_csv('chords_all_songs.csv')
hook_data = hook_data.drop(columns='Unnamed: 0')
hook_data.head()

# Display size of dataframe
len(hook_data)

# Import data retrieved from last.fm and spotify for unique songs subset of HookTheory data
stats = pd.read_csv('spotify_data.csv')
stats.head()

# Rename first column as 'artist'
stats = stats.rename(columns={'0':'artist'})

# Create a truncated dataframe
lastfm_raw = stats[['artist',
                       'song',
                       'listener count (last.fm)',
                       'play count (last.fm)',
                       'genre (last.fm)']]

# Drop null entries from dataframe
lastfm_raw = lastfm_raw.dropna(how='any')
lastfm_raw.head()

# Separate genre data for each song into individual rows
lastfm_genre = pd.DataFrame(lastfm_raw['genre (last.fm)'].str.split(', ').tolist(),index=lastfm_raw['song']).stack()
lastfm_genre

# Reset index to resulting dataframe
lastfm_genre = lastfm_genre.reset_index([0,'song'])

# Assign column headers
lastfm_genre.columns = ['song','genre (last.fm)']
lastfm_genre.head()

# Drop original genre column from original dataframe and merge with new dataframe
lastfm_raw = lastfm_raw.drop(columns=['genre (last.fm)'])
lastfm_stats = pd.merge(lastfm_raw, lastfm_genre, how='outer', on='song')
lastfm_stats.head()

# Determine if the values for artist and genre for each row match
# if lastfm_stats.loc[3,'artist'] == lastfm_stats.loc[3,'genre (last.fm)']:
#     print('True')
# else:
#     print('False')

# Inititalize list of index information for rows to be dropped from dataframe
droplist = []

# Use for loop to determine if either the artist name or the song name matches the genre for each row
# If either condition is met, add index number to droplist list
for song in range(len(lastfm_stats)):
    artist = lastfm_stats.loc[song,'artist']
    track = lastfm_stats.loc[song,'song']
    genre = lastfm_stats.loc[song,'genre (last.fm)']
    if (artist == genre) | (track == genre):
        droplist.append(song)

# Drop dataframe entries based on index information contained in droplist list
lastfm_stats.drop(droplist,inplace=True)
lastfm_stats.head()

# Create a truncated dataframe
spotify_stats = stats[['artist',
                       'song',
                       'popularity (Spotify)',
                       'release date (Spotify)']]

# Drop null entries from dataframe
spotify_stats = spotify_stats.dropna(how='any')
spotify_stats.head()

# Display size of dataframe
len(stats)

# Convert release date entries to datetime type and format to show only year
spotify_stats['release date (Spotify)'] = pd.to_datetime(spotify_stats['release date (Spotify)'])
spotify_stats['release date (Spotify)'] = pd.DatetimeIndex(spotify_stats['release date (Spotify)']).year
spotify_stats.head()

# Merge full HookTheory data set with last.fm data
combined_lastfm = pd.merge(hook_data, lastfm_stats, how='left', on=['artist','song'])
combined_lastfm.head()

# Display size of dataframe
len(combined_lastfm)

# Drop duplicate entries from the merged dataframe using values from specific columns as criteria 
combined_lastfm = combined_lastfm.drop_duplicates(subset=['artist','cp','song','genre (last.fm)'])

# Cast listener count and play count data as integers
# combined_lastfm['listener count (last.fm)'] = combined_lastfm['listener count (last.fm)'].astype(int)
# combined_lastfm['play count (last.fm)'] = combined_lastfm['play count (last.fm)'].astype(int)

combined_lastfm.head()

# Display size of dataframe
len(combined_lastfm)

# Save resulting dataframe as a csv file
combined_lastfm.to_csv('combined_hook-lastfm.csv')

# Merge full HookTheory data set with Spotify data
combined_spotify = pd.merge(hook_data, spotify_stats, how='left', on=['artist','song'])

# Drop null entries from merged dataframe
combined_spotify = combined_spotify.dropna(how='any')

# Cast popularity and release data as integers
combined_spotify['release date (Spotify)'] = combined_spotify['release date (Spotify)'].astype(int)
combined_spotify['popularity (Spotify)'] = combined_spotify['popularity (Spotify)'].astype(int)

combined_spotify.head()

# Display size of dataframe
len(combined_spotify)

# Save resulting dataframe as a csv file
combined_spotify.to_csv('combined_hook-spotify.csv')

# Drop duplicate entries from the merged dataframe using values from specific columns as criteria 
unique_spotify = combined_spotify.drop_duplicates(subset=['artist','song'])
unique_spotify.head()

# Create a truncated dataframe
sub_spotify = unique_spotify[['artist','song','popularity (Spotify)','release date (Spotify)']]
sub_spotify.head()

combined_all = pd.merge(combined_lastfm, sub_spotify, how='left', on=['artist','song'])
combined_all = combined_all.dropna(how='any')
combined_all['listener count (last.fm)'] = combined_all['listener count (last.fm)'].astype(int)
combined_all['play count (last.fm)'] = combined_all['play count (last.fm)'].astype(int)
combined_all['release date (Spotify)'] = combined_all['release date (Spotify)'].astype(int)
combined_all['popularity (Spotify)'] = combined_all['popularity (Spotify)'].astype(int)
combined_all.head()

# Display size of dataframe
len(combined_all)

# Save resulting dataframe as a csv file
combined_all.to_csv('combined_all.csv')

# Create a dataframe containing only one entry for each song 
combined_unique = combined_all.drop_duplicates(subset=['artist','song'])
combined_unique.head()

# Save resulting dataframe as a csv file
combined_unique.to_csv('combined_unique.csv')
